refactor(auth): 重构身份验证和工作区查询逻辑
- 新增 createSessionClient 函数以创建带有会话的 Appwrite 客户端
- 将 auth 和 workspaces 相关的 actions重命名为 queries
- 更新相关页面以使用新的查询函数
- 优化代码结构，提高可维护性和可重用性

timestamp - 7:12:03
