+++ src/app/(auth)/sign-in/page.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import SignInCard from '@/features/auth/components/sign-in-card';
-import { getCurrent } from '@/features/auth/actions';
+import { getCurrent } from '@/features/auth/queries';
 import { redirect } from 'next/navigation';
 
 const SignInPage = async () => {
