+++ src/app/(auth)/sign-up/page.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import SignUpCard from '@/features/auth/components/sign-up-card';
-import { getCurrent } from '@/features/auth/actions';
+import { getCurrent } from '@/features/auth/queries';
 import { redirect } from 'next/navigation';
 
 const SignUpPage = async () => {
