+++ src/app/(dashboard)/page.tsx
@@ -1,8 +1,8 @@
-import { getCurrent } from '@/features/auth/actions';
+import { getCurrent } from '@/features/auth/queries';
 import { redirect } from 'next/navigation';
-import { getWorkspaces } from '@/features/workspaces/actions'
+import { getWorkspaces } from '@/features/workspaces/queries';
 
-export default async function Home() {
+export default async function Home () {
   const user = await getCurrent();
 
   if (!user) {
