+++ src/app/(dashboard)/workspaces/[workspaceId]/page.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
-import { getCurrent } from '@/features/auth/actions';
+import { getCurrent } from '@/features/auth/queries';
 import { redirect } from 'next/navigation';
 
 const WorkspaceIdPage = async () => {
   const user = await getCurrent();
-  
+
   if (!user) {
     redirect('/sign-in');
   }
