+++ src/app/(standalone)/workspaces/[workspaceId]/settings/page.tsx
@@ -1,9 +1,9 @@
-import { getCurrent } from '@/features/auth/actions';
+import { getCurrent } from '@/features/auth/queries';
 import { redirect } from 'next/navigation';
 import {
   EditWorkspaceForm,
 } from '@/features/workspaces/components/edit-workspace-form';
-import { getWorkspace } from '@/features/workspaces/actions';
+import { getWorkspace } from '@/features/workspaces/queries';
 
 interface WorkspaceIdSettingsPageProps {
   params: {
