+++ src/app/(standalone)/workspaces/create/page.tsx
@@ -1,7 +1,7 @@
 import {
   CreateWorkspaceForm,
 } from '@/features/workspaces/components/create-workspace-form';
-import { getCurrent } from '@/features/auth/actions';
+import { getCurrent } from '@/features/auth/queries';
 import { redirect } from 'next/navigation';
 
 const WorkspaceCreatePage = async () => {
