+++ src/features/auth/queries.ts
@@ -1,0 +1,12 @@
+import { createSessionClient } from '@/lib/appwrite';
+
+const getCurrent = async () => {
+  try {
+    const { account } = await createSessionClient();
+    return await account.get();
+  } catch {
+    return null;
+  }
+};
+
+export { getCurrent };
