+++ src/features/workspaces/queries.ts
@@ -1,9 +1,8 @@
-import { Account, Client, Databases, Query } from 'node-appwrite';
-import { cookies } from 'next/headers';
-import { AUTH_COOKIE } from '@/features/auth/constants';
+import { Query } from 'node-appwrite';
 import { DATABASE_ID, MEMBERS_ID, WORKSPACES_ID } from '@/config';
 import { getMember } from '@/features/members/utils';
 import { Workspace } from '@/features/workspaces/types';
+import { createSessionClient } from '@/lib/appwrite';
 
 interface GetWorkspaceProps {
   workspaceId: string;
@@ -11,22 +10,8 @@ interface GetWorkspaceProps {
 
 const getWorkspaces = async () => {
   try {
-    const client = new Client().setEndpoint(
-      process.env.NEXT_PUBLIC_APPWRITE_ENDPOINT!).
-      setProject(process.env.NEXT_PUBLIC_APPWRITE_PROJECT!);
-
-    const session = cookies().get(AUTH_COOKIE);
-
-    if (!session) {
-      return { documents: [], total: 0 };
-    }
-
-    client.setSession(session.value);
-    const databases = new Databases(client);
-    const account = new Account(client);
+    const { account, databases } = await createSessionClient();
     const user = await account.get();
-
-    // 获取当前用户的工作区
     const members = await databases.listDocuments(
       DATABASE_ID,
       MEMBERS_ID,
@@ -54,19 +39,7 @@ const getWorkspaces = async () => {
 
 const getWorkspace = async ({ workspaceId }: GetWorkspaceProps) => {
   try {
-    const client = new Client().setEndpoint(
-      process.env.NEXT_PUBLIC_APPWRITE_ENDPOINT!).
-      setProject(process.env.NEXT_PUBLIC_APPWRITE_PROJECT!);
-
-    const session = cookies().get(AUTH_COOKIE);
-
-    if (!session) {
-      return null;
-    }
-
-    client.setSession(session.value);
-
-    const [databases, account] = [new Databases(client), new Account(client)];
+    const { account, databases } = await createSessionClient();
     const user = await account.get();
     const member = await getMember({
       databases,
